fix tests
Copier's run_copy on a template, version controlled with git, by default searches for the latest tag conforming to PEP-440 (https://peps.python.org/pep-0440/#public-version-identifiers) and uses that to generate. This meant that it was still using the 0.4.0 version. We now set it to HEAD explicitly in the tests.

Note that we didn't see this failure before in the separate repo because there was no history there.