Remove docs from minimum profile
Introduces a question AddLocalDocumentation to control the existence of
the `docs` folder.

If the user selects "Online documentation", then "Local documentation"
is implied. The logic behind this approach is simply to define the
computed variable `AddLocalDocumentation` based on either flag.
The downside is that the generated answers will still only show the
online flag.

Related to #593